refactor: move ingestion/utils under utils/utils (#35)
- Fixed imports under tests from mnemolet_core to mnemolet
- Moved hash fn from ingestion/utils to utils/utils
